Fix check for mouse click event handling
Corrected logic that tested if a click event is being handled by
simplifing the check for "None" and empty string and exiting the
function if either is true.

It turns out that when a script executes due to i3blocks events,
the BLOCK_BUTTON value exists but it is set to empty string.